requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.0
groq==0.3.0
pydantic==2.6.1
psycopg2-binary==2.9.9
psycopg2